pyproject.toml
[project]
name = "hivemind-shared-compute"
version = "0.1.0"
description = "Democratizing high-end computing by pooling idle CPU/GPU resources across devices"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "psutil>=5.9",
    "pydantic>=2.0",
    "pydantic-settings",
    "rich",
    "zeroconf",
    "pyzmq",
    "structlog",
]

[project.optional-dependencies]
gpu = ["torch"]  # CUDA on Windows, MPS on Mac Studio
dev = [
    "pytest",
    "black",
    "ruff",
    "mypy",
]
ray = ["ray[default]>=2.0"]

[project.scripts]
hivemind-host = "hivemind.host:main"
hivemind-client = "hivemind.client:main"
hivemind = "hivemind.client.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/hivemind"]

[tool.hatch.build.targets.editable]
mode = "in-tree"